Fixed all the settings being erased
That was a major issue, glad it's resolved.
