+++ languages.xml
@@ -1,2 +1,56 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<languages />
+<languages>
+  <language name="C" index="0">
+    <extensions>
+      <extension text="c" />
+      <extension text="h" />
+    </extensions>
+    <lineCommentChar text="//" />
+    <blockCommentChars opening="/*" closing="*/" />
+  </language>
+  <language name="C++" index="1">
+    <extensions>
+      <extension text="cc" />
+      <extension text="cpp" />
+      <extension text="cxx" />
+      <extension text="C" />
+      <extension text="c++" />
+      <extension text="h" />
+      <extension text="hh" />
+      <extension text="hpp" />
+      <extension text="hxx" />
+      <extension text="h++" />
+    </extensions>
+    <lineCommentChar text="//" />
+    <blockCommentChars opening="/*" closing="*/" />
+  </language>
+  <language name="C#" index="2">
+    <extensions>
+      <extension text="cs" />
+    </extensions>
+    <lineCommentChar text="//" />
+    <blockCommentChars opening="/*" closing="*/" />
+  </language>
+  <language name="Java" index="3">
+    <extensions>
+      <extension text="java" />
+    </extensions>
+    <lineCommentChar text="//" />
+    <blockCommentChars opening="/*" closing="*/" />
+  </language>
+  <language name="PHP" index="4">
+    <extensions>
+      <extension text="php" />
+    </extensions>
+    <lineCommentChar text="//" />
+    <lineCommentChar text="#" />
+    <blockCommentChars opening="/*" closing="*/" />
+  </language>
+  <language name="Python" index="5">
+    <extensions>
+      <extension text="py" />
+    </extensions>
+    <lineCommentChar text="#" />
+    <blockCommentChars opening="&quot;&quot;&quot;" closing="&quot;&quot;&quot;" />
+  </language>
+</languages>
