+++ src/linecounter/ProjectReader.java
@@ -1,7 +1,6 @@
 package linecounter;
 
 import java.io.File;
-import javax.swing.JFileChooser;
 
 /**
  * ProjectReader will take a given folder, search through it, and use a SourceFileReader to process
@@ -19,10 +18,6 @@ public class ProjectReader {
 
     private long totalLines     = 0;
     private long commentLines   = 0;
-
-    public File rootFolder = null;
-    private Language language;
-    private boolean ignoreComments;
     
     public Settings settings;
 
@@ -71,7 +66,9 @@ public class ProjectReader {
                 numFiles++;
 
                 // Create a source file reader for the file, and add its line count to the total count
-                SourceFileReader reader =  new SourceFileReader(file, language, ignoreComments);
+                SourceFileReader reader =  new SourceFileReader(file, 
+                		settings.selectedLang(), 
+                		settings.ignoreComments);
                 totalLines += reader.totalLines;
                 commentLines += reader.commentLines;
             }
@@ -97,48 +94,12 @@ public class ProjectReader {
         
         return false;
     }
-
-    /**
-     * Choose the root folder of a project using a JFileChooser dialog.
-     * @return A file on the system somewhere, to be treated as the root folder of a project.
-     * @throws NoFileChosenException If no file is chosen (i.e. the user hits the "cancel" button)
-     */
-    public void chooseRootFolder() throws NoFileChosenException{
-        // Create a new file chooser dialog
-        JFileChooser chooser = new JFileChooser();
-        chooser.setFileSelectionMode(JFileChooser.DIRECTORIES_ONLY);
-
-        // If the user hits the cancel button, throw an exception
-        if(chooser.showOpenDialog(null)!=JFileChooser.APPROVE_OPTION)
-            throw new NoFileChosenException();
-
-        this.rootFolder = chooser.getSelectedFile();
-    }
-
-    /**
-     * Set the language index for this project.
-     * @param languageIndex
-     */
-    public void setLanguage(int languageIndex) {
-        settings.selectLanguage(languageIndex);
-        this.language = settings.selectedLang();
-    }
     
-    /**
-     * Set whether to ignore comments or not.
-     * @param ignoreComments 
-     */
-    public void setIgnoreComments(boolean ignoreComments) {
-    	this.ignoreComments = ignoreComments;
-    }
-
     /**
      * Create a new ProjectReader object with the given settings.
      * @param settings
      */
     public ProjectReader(Settings settings) {
     	this.settings = settings;
-        this.ignoreComments = settings.ignoreComments;
-        this.language = settings.selectedLang();
     }
 }
